backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
aiohttp==3.9.5
beautifulsoup4==4.12.3
aiosqlite==0.20.0
python-dotenv==1.0.1
